backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
numpy>=1.26.0
faiss-cpu>=1.8.0
openai>=1.40.0
pydantic>=2.7.0
python-dotenv>=1.0.1
sentence-transformers>=3.0.0
